refactor(order): 重构订单结算流程与金额分单位存储
- 实现全站金额字段从元(decimal)改为分(int)存储，消除浮点精度问题
- 重构订单预览和提交流程，补全Contract至DTO的DDD合规链路
- 精简OrderCheckoutTransformer，移除冗余的toCent转换和手动字段拼装
- 修复payByBalance支付方式覆盖bug和库存扣减顺序问题
- 添加策略工厂自动注册机制和OrderCreatedEvent监听器
- 更新前端价格展示适配，统一使用分单位进行金额计算
- 完善价格校验逻辑，确保前后端金额一致性验证

+++ app/Domain/Coupon/Entity/CouponEntity.php
@@ -26,9 +26,9 @@ final class CouponEntity
 
     private ?string $type;
 
-    private ?float $value;
+    private ?int $value;
 
-    private ?float $minAmount;
+    private ?int $minAmount;
 
     private ?int $totalQuantity;
 
@@ -148,13 +148,13 @@ final class CouponEntity
         return $this;
     }
 
-    public function setValue(?float $value): self
+    public function setValue(?int $value): self
     {
         $this->value = $value;
         return $this;
     }
 
-    public function setMinAmount(?float $minAmount): self
+    public function setMinAmount(?int $minAmount): self
     {
         $this->minAmount = $minAmount;
         return $this;
@@ -325,12 +325,12 @@ final class CouponEntity
         return $this->type;
     }
 
-    public function getValue(): ?float
+    public function getValue(): ?int
     {
         return $this->value;
     }
 
-    public function getMinAmount(): ?float
+    public function getMinAmount(): ?int
     {
         return $this->minAmount;
     }
