+++ app/Domain/GroupBuy/Contract/GroupBuyCreateInput.php
@@ -25,9 +25,9 @@ interface GroupBuyCreateInput
 
     public function getSkuId(): int;
 
-    public function getOriginalPrice(): float;
+    public function getOriginalPrice(): int;
 
-    public function getGroupPrice(): float;
+    public function getGroupPrice(): int;
 
     public function getMinPeople(): int;
